Move interfaces to its own folder
