+++ src/Ether.Network/Interfaces/INetConnection.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Net.Sockets;
+
+namespace Ether.Network.Interfaces
+{
+    public interface INetConnection
+    {
+        Guid Id { get; }
+
+        Socket Socket { get; }
+    }
+}
