+++ src/Ether.Network/Interfaces/INetUser.cs
@@ -1,0 +1,9 @@
+﻿namespace Ether.Network.Interfaces
+{
+    public interface INetUser : INetConnection
+    {
+        void Send();
+
+        void HandleMessage();
+    }
+}
