+++ src/Ether.Network/Packets/NetPacketStream.cs
@@ -1,4 +1,4 @@
-﻿using Ether.Network.Core;
+﻿using Ether.Network.Interfaces;
 using System;
 using System.IO;
 using System.Linq;
