+++ src/Ether.Network/Utils/SocketAsyncUtils.cs
@@ -1,4 +1,4 @@
-﻿using Ether.Network.Core;
+﻿using Ether.Network.Interfaces;
 using System;
 using System.Net.Sockets;
