+++ test/Ether.Network.Tests/NetPacketStreamTest.cs
@@ -1,4 +1,4 @@
-﻿using Ether.Network.Core;
+﻿using Ether.Network.Interfaces;
 using Ether.Network.Packets;
 using Ether.Network.Tests.Helpers;
 using System;
